logmpx: make sure the LogMessage is write protected while being dispatched
Write protection ensures that parallal branches along the log processing
pipeline receive the same message. The log multiplexer plays an important
role as that's the place where this parallelism is implemented.

Cloning of messages happen automatically, when there's a write protected
message and still it needs to be modified.

Write protection is enabled if a message is enqueued to a destination.
So with all this said, in most cases syslog-ng behaved properly, as long
as the the branch delivered the message to a queue.

As it does not always happen, it makes sense to "write_protect" early, during
message dispatch.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>