logmatcher: don't use indirect values if input ref_handle is a match
As we are setting matches in bulk, it might happen that the input
is actually a match in itself, being overwritten by our
feed_backrefs() loop.

LogMessage handles self-references, however the case might be a more subtle:

* the input is $1
* the regexp is setting $0, $1 and $2

$2 would be set indirectly, as $2 is not the same as $1, however we just
overwrote $1 in the previous iteration.

match("(.*):(.*)" value('1'));

When we match, the first subpattern would be set into $1, which also
happens to be the input.

On the performance impact, matches change frequently anyway, so
setting indirect values and then handling the change of the match
is probably slower anyway.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>